Chapter Questions Question three - I have reviewed it and changed it to integers. Pushed for review

+++ src/week3/Dollars.java
@@ -5,14 +5,14 @@ import java.util.Scanner;
 public class Dollars {
     public static void dollars(){
         Scanner input = new Scanner(System.in);
-        Double totalDollars = input.nextDouble();
-        Double twenties = totalDollars / 20;
-        Double leftOver = totalDollars % 20;
-        Double tens = leftOver / 10;
+        int totalDollars = input.nextInt();
+        int twenties = totalDollars / 20;
+        int leftOver = totalDollars % 20;
+        int tens = leftOver / 10;
         leftOver = leftOver % 10;
-        Double fives = leftOver / 5;
+        int fives = leftOver / 5;
         leftOver = leftOver % 5;
-        Double ones = leftOver / 1;
+        int ones = leftOver / 1;
         leftOver = leftOver % 1;
         System.out.println("You have - " +
                 twenties + " x twenties. " +
